Cleaned up model imports in EditableDisplay; extracted typesToComponentMap from contentItems EditableDisplay and moved it into a separate file; removed classname props in favor of data-test-ids.
